// web/src/utils/backend-client/types/accessRuleStatus.ts
/**
 * Generated by orval v6.8.1 🍺
 * Do not edit manually.
 * Approvals
 * Granted Approvals API
 * OpenAPI spec version: 1.0
 */

/**
 * The status of an Access Rule.
 */
export type AccessRuleStatus = typeof AccessRuleStatus[keyof typeof AccessRuleStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const AccessRuleStatus = {
  ACTIVE: 'ACTIVE',
  ARCHIVED: 'ARCHIVED',
} as const;